// Home/transaction/fin-02-invoice/model/fin03-clinic.ts
export class Fin03Clinic {
    clinicName: string;
    clinicCode: string;
    fin08bEquipmentCount: number = 0;
    fin08bMaintenanceCharge: number = 0.0;
    fin08cEquipmentCount: number = 0;
    fin08cMaintenanceCharge: number = 0.0;
    fin08MaintenanceCharge: number = 0.0;
    fin09PenaltyCharge: any;
    fin09PenaltyCharges:any;
    netMaintenanceCharge: number;
}